requirements.txt
# Kalshi NBA In-Game Trading Analysis
# Python Dependencies

# Core data science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Machine learning
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

